Using static call to Await.await as the instrumentation point. Runtime instrumentation working during the test cases.
