+++ orbit-async/src/test/java/com/ea/orbit/async/test/JoinTest.java
@@ -26,46 +26,55 @@
  THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.ea.orbit.async.test.manual;
+package com.ea.orbit.async.test;
 
 import com.ea.orbit.async.Async;
-import com.ea.orbit.async.instrumentation.InstrumentAsync;
+import com.ea.orbit.async.Await;
 
 import org.junit.Test;
 
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 import java.util.concurrent.CompletableFuture;
 
+import static com.ea.orbit.async.Await.await;
 import static org.junit.Assert.assertEquals;
 
-public class StaticTest
+public class JoinTest
 {
-
-    public static class StaticUse
+    static
     {
-        static String concat(int i, long j, float f, double d, Object obj, boolean b)
+        Await.init();
+    }
+
+    public static class OtherJoinCalls
+    {
+        @Async
+        public CompletableFuture<Object> doSomething(CompletableFuture<String> blocker)
         {
-            return i + ":" + j + ":" + f + ":" + d + ":" + obj + ":" + b;
+            int local = 7;
+            String res = ":" + Math.max(local, await(blocker).length()) + ":" + join() + ":" + join(2, 3);
+            return CompletableFuture.completedFuture(res);
         }
 
-        @Async
-        public static CompletableFuture<Object> staticMethod(CompletableFuture<String> blocker, int var)
+        public Object join()
         {
-            return CompletableFuture.completedFuture(concat(var, 10_000_000_000L, 1.5f, 3.5d, blocker.join(), true));
+            return 9;
+        }
+
+        public Object join(int a, int b)
+        {
+            return a + b;
         }
     }
 
     @Test
-    public void testStaticMethodWithPrimitives() throws IllegalAccessException, InstantiationException, InvocationTargetException, NoSuchMethodException
+    public void testOtherJoinMethods() throws IllegalAccessException, InstantiationException
     {
-        InstrumentAsync ins = new InstrumentAsync();
-        Class<?> newClass = ins.instrument(StaticUse.class);
-        final Method method = newClass.getMethod("staticMethod", CompletableFuture.class, int.class);
+        OtherJoinCalls a = new OtherJoinCalls();
 
         CompletableFuture<String> blocker = new CompletableFuture<>();
-        final CompletableFuture<String> res = (CompletableFuture<String>) method.invoke(null, blocker, 5);
-        blocker.complete("zzz");
-        assertEquals("5:10000000000:1.5:3.5:zzz:true", res.join());
+        final CompletableFuture<Object> res = a.doSomething(blocker);
+        blocker.complete("0123456789");
+        assertEquals(":10:9:5", res.join());
     }
+
 }
