+++ orbit-async/src/test/java/com/ea/orbit/async/test/RuntimeInstrTest.java
@@ -29,20 +29,21 @@
 package com.ea.orbit.async.test;
 
 import com.ea.orbit.async.Async;
-import com.ea.orbit.async.instrumentation.InitializeAsync;
+import com.ea.orbit.async.Await;
 
 import org.junit.Test;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.concurrent.CompletableFuture;
 
+import static com.ea.orbit.async.Await.await;
 import static org.junit.Assert.assertEquals;
 
 public class RuntimeInstrTest
 {
     static
     {
-        new InitializeAsync();
+        Await.init();
     }
 
     private static class Basic
@@ -55,7 +56,7 @@ public class RuntimeInstrTest
         @Async
         public CompletableFuture<String> doSomething(CompletableFuture<String> blocker, int var)
         {
-            return CompletableFuture.completedFuture(concat(var, 10_000_000_000L, 1.5f, 3.5d, blocker.join(), true));
+            return CompletableFuture.completedFuture(concat(var, 10_000_000_000L, 1.5f, 3.5d, await(blocker), true));
         }
     }
